+++ pages/about.php
@@ -33,7 +33,7 @@ THE SOFTWARE.</p></pre>
     <td>&nbsp;</td>
   </tr>
   <tr>
-    <td><p>Version: <strong><?php echo $ver;?></strong></p>
+    <td><p>Version: <strong><?php echo file_get_contents('../VERSION.md');?></strong></p>
     <p><a href="https://github.com/globaldyne/parfumvault" target="_blank"><strong>Github Repo</strong></a></p>
     <p>Theme by <a href="https://startbootstrap.com/themes/sb-admin-2/" target="_blank"><strong>SB Admin 2</strong></a></p></td>