+++ pages/getFormMeta.php
@@ -1,3 +1,12 @@
+<?php
+require_once('../inc/config.php');
+require_once('../inc/opendb.php');
+require_once('../inc/settings.php');
+
+if($_GET['id']){
+	$id = mysqli_real_escape_string($conn, $_GET['id']);
+	$info = mysqli_fetch_array(mysqli_query($conn, "SELECT * FROM formulasMetaData WHERE id = '$id'"));
+?>
 <link href="../css/sb-admin-2.css" rel="stylesheet">
 <link href="../css/bootstrap-select.min.css" rel="stylesheet">
 <link href="../css/bootstrap-editable.css" rel="stylesheet">
@@ -9,7 +18,7 @@
   
 <script src="../js/bootstrap-select.js"></script>
 <script src="../js/bootstrap-editable.js"></script>
-  
+
 
 <style>
 .form-inline .form-control {
@@ -19,15 +28,7 @@
 }
 </style>
 
-<?php
-require_once('../inc/config.php');
-require_once('../inc/opendb.php');
-require_once('../inc/settings.php');
 
-if($_GET['id']){
-	$id = mysqli_real_escape_string($conn, $_GET['id']);
-	$info = mysqli_fetch_array(mysqli_query($conn, "SELECT * FROM formulasMetaData WHERE id = '$id'"));
-?>
 
 <table class="table table-bordered" id="formula_metadata" cellspacing="0">
   <tr>
@@ -36,6 +37,10 @@ if($_GET['id']){
   <tr>
     <td width="20%">Created:</td>
     <td width="80%"><?php echo $info['created'];?></td>
+  </tr>
+  <tr>
+    <td>Profile:</td>
+    <td><a href="#" id="profile" data-type="select" data-pk="profile" data-title="<?php echo $info['profile'];?>"></a></td>
   </tr>
   <tr>
     <td>Notes:</td>
@@ -62,9 +67,22 @@ $(document).ready(function(){
   mode: 'inline',
   dataType: 'json',
       success: function(response, newValue) {
-        if(response.status == 'error') return response.msg; //msg will be shown in editable form
+        if(response.status == 'error') return response.msg; 
     },
 
  });
+  
+  $('#profile').editable({
+	value: "<?php echo $info['profile'];?>",
+  	title: 'Profile',
+  	url: "/pages/update_data.php?formulaMeta=<?php echo $info['name']; ?>",
+    source: [
+             {value: 'oriental', text: 'Oriental'},
+             {value: 'woody', text: 'Woody'},
+             {value: 'floral', text: 'Floral'},
+             {value: 'fresh', text: 'Fresh'}
+          ]
+    });
+  
   })
 </script>